+++ CHANGELOG.md
@@ -5,6 +5,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+### Changed
+
+- replace all "control-group" with "control-container" in the CSS and HTML templates for better semantics and to allow more flexible styling. The class names for control groups are now "control-container control-container--slider", "control-container control-container--color-modern", and "control-container control-container--color-standard".
+
+### Added
+
+- Add option to group variables in the UI under custom headings. If a property in the config object contains nested properties instead of a `type`, it will be treated as a group heading and the nested properties will be rendered under it. 
+
 ## [1.5.1] - 2026-02-10
 
 ### Changed